ilmol/requirements.txt
Pillow==5.2.0
-e git://github.com/jamesacampbell/iptcinfo3.git#egg=iptcinfo3